+++ elaston/__init__.py
@@ -1,3 +1,4 @@
 from . import _version
 
+
 __version__ = _version.get_versions()["version"]
